Add check for existing bashio before copy
Check if bashio is already installed before copying.

+++ .templates/00-banner.sh
@@ -16,7 +16,9 @@ if ! bashio::supervisor.ping 2>/dev/null; then
     bashio::log.blue "Version : ${BUILD_VERSION:-1.0}"
     bashio::log.blue "Config source: ENV + /data/options.json"
     bashio::log.blue '-----------------------------------------------------------'
-    cp -rf /usr/local/lib/bashio-standalone.sh /usr/bin/bashio
+    if [ ! -f /usr/bin/bashio ]; then
+        cp -rf /usr/local/lib/bashio-standalone.sh /usr/bin/bashio
+    fi
     grep -rl "^#!.*bashio" /etc |
     while IFS= read -r f; do
         grep -qF "source /usr/local/lib/bashio-standalone.sh" "$f" && continue
